Create `addParams`, replacing `queryEncoder`
Summary:
The existing `queryEncoder` function had a significant flaw: it always
returned a string starting with `"?"`, but the initial delimiter should
be `"&"` if the URL already contains a query component. This could
happen unexpectedly if, say, an application embeds TensorBoard and
monkey-patches the router to include some parameters in every URL, so
that a call like `pluginRoute` has query parameters where the original
author of the plugin expects none.

A simple solution is to replace this with a function `addParams` that
constructs the query string and attaches it to the URL in one go, adding
to an existing string if one exists. This new function also now supports
repeated parameters for convenience.

Test Plan:
I tested that all existing plugins work as intended. I didn’t test the
JS test because we have no way to run them.

wchargin-branch: add-params